Move and refactor spotify client
Changes to request user data
